rework trail text visibility logic
